implement methods in base class instead of generating
